融合

+++ src/processdata_pkg/src/HungarianAlgorithm.cpp
@@ -5,10 +5,10 @@
 #include <algorithm>
 
 // Point结构实现
-Point::Point(double _x, double _y) : x(_x), y(_y) {}
+Point::Point(float _x, float _y) : x(_x), y(_y) {}
 
 // 匈牙利算法实现
-HungarianAlgorithm::HungarianAlgorithm(const std::vector<std::vector<double>>& cost_matrix)
+HungarianAlgorithm::HungarianAlgorithm(const std::vector<std::vector<float>>& cost_matrix)
         : cost(cost_matrix), N(cost_matrix.size()), M(cost_matrix.empty() ? 0 : cost_matrix[0].size()) {
     u.assign(N + 1, 0);
     v.assign(M + 1, 0);
@@ -26,17 +26,17 @@ std::vector<int> HungarianAlgorithm::solve() {
     for (int i = 1; i <= N; ++i) {
         p[0] = i;
         int j0 = 0;
-        std::vector<double> minv(M + 1, std::numeric_limits<double>::max());
+        std::vector<float> minv(M + 1, std::numeric_limits<float>::max());
         std::vector<bool> used(M + 1, false);
 
         do {
             used[j0] = true;
             int i0 = p[j0], j1;
-            double delta = std::numeric_limits<double>::max();
+            float delta = std::numeric_limits<float>::max();
 
             for (int j = 1; j <= M; ++j) {
                 if (!used[j]) {
-                    double cur = cost[i0 - 1][j - 1] - u[i0] - v[j];
+                    float cur = cost[i0 - 1][j - 1] - u[i0] - v[j];
                     if (cur < minv[j]) {
                         minv[j] = cur;
                         way[j] = j0;
@@ -77,7 +77,7 @@ std::vector<int> HungarianAlgorithm::solve() {
 }
 
 // 计算欧氏距离
-double euclideanDistance(const Point& p1, const Point& p2) {
+float euclideanDistance(const Point& p1, const Point& p2) {
     return std::sqrt(std::pow(p1.x - p2.x, 2) + std::pow(p1.y - p2.y, 2));
 }
 
@@ -85,7 +85,7 @@ double euclideanDistance(const Point& p1, const Point& p2) {
 std::vector<std::pair<int, int>> associatePoints(
         const std::vector<Point>& sensor1_points,
         const std::vector<Point>& sensor2_points,
-        double max_distance) {
+        float max_distance) {
 
     std::vector<std::pair<int, int>> matches;
 
@@ -94,13 +94,13 @@ std::vector<std::pair<int, int>> associatePoints(
     }
 
     // 创建成本矩阵
-    std::vector<std::vector<double>> cost_matrix(
+    std::vector<std::vector<float>> cost_matrix(
             sensor1_points.size(),
-            std::vector<double>(sensor2_points.size(), max_distance * 2));  // 初始化为最大距离的2倍
+            std::vector<float>(sensor2_points.size(), max_distance * 2));  // 初始化为最大距离的2倍
 
     for (size_t i = 0; i < sensor1_points.size(); ++i) {
         for (size_t j = 0; j < sensor2_points.size(); ++j) {
-            double dist = euclideanDistance(sensor1_points[i], sensor2_points[j]);
+            float dist = euclideanDistance(sensor1_points[i], sensor2_points[j]);
             if (dist <= max_distance) {
                 cost_matrix[i][j] = dist;
             }
